@@ -1,5 +1,6 @@
 {
   "name": "react-petstore",
+  "type": "module",
   "version": "1.0.0",
   "description": "A skeleton react app for the petstore api.",
   "keywords": [
@@ -12,43 +13,49 @@
   "repository": "chubbyts/react-petstore",
   "scripts": {
     "build": "tsc && vite build",
-    "cs-fix": "prettier --write src",
-    "cs": "prettier --check src",
+    "cs-fix": "prettier --write src tests *.cjs *.js *.ts",
+    "cs": "prettier --check src tests *.cjs *.js *.ts",
     "develop": "vite",
-    "start": "npm i && npm run develop",
+    "lint-fix": "eslint src tests *.cjs *.js *.ts --fix",
+    "lint": "eslint src tests *.cjs *.js *.ts",
+    "start": "pnpm i && pnpm cs-fix && pnpm develop",
     "test": "vitest"
   },
   "dependencies": {
     "@chubbyts/chubbyts-throwable-to-error": "^1.2.1",
-    "autoprefixer": "^10.4.17",
+    "autoprefixer": "^10.4.18",
+    "cross-fetch": "^4.0.0",
     "date-fns": "^3.3.1",
-    "history": "^5.3.0",
     "postcss": "^8.4.35",
-    "qs": "^6.11.2",
+    "qs": "^6.12.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-hook-form": "^7.50.1",
-    "react-router-dom": "^6.22.1",
+    "react-router-dom": "^6.22.3",
     "tailwindcss": "^3.4.1",
     "zod": "^3.22.4"
   },
   "devDependencies": {
+    "@chubbyts/chubbyts-function-mock": "^1.4.1",
     "@testing-library/jest-dom": "^6.4.2",
     "@testing-library/react": "^14.2.1",
     "@testing-library/user-event": "^14.5.2",
-    "@types/node": "^20.11.20",
+    "@types/node": "^20.11.25",
     "@types/prettier": "^2.7.3",
     "@types/qs": "^6.9.12",
-    "@types/react": "^18.2.60",
-    "@types/react-dom": "^18.2.19",
+    "@types/react": "^18.2.64",
+    "@types/react-dom": "^18.2.21",
+    "@typescript-eslint/eslint-plugin": "^7.1.0",
+    "@typescript-eslint/parser": "^7.1.0",
     "@vitejs/plugin-react": "^4.2.1",
     "@vitest/coverage-v8": "^1.3.1",
-    "cross-fetch": "^4.0.0",
-    "fetch-mock": "^9.11.0",
+    "eslint": "^8.57.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-react": "^7.34.0",
+    "eslint-plugin-tailwindcss": "^3.14.3",
     "jsdom": "^24.0.0",
     "prettier": "^2.8.8",
-    "typescript": "^5.3.3",
-    "vite": "^5.1.4",
+    "typescript": "^5.4.2",
+    "vite": "^5.1.5",
     "vitest": "^1.3.1"
   },
   "engines": {
