@@ -1,7 +1,7 @@
-module.exports = {
+export default {
   plugins: {
     'tailwindcss/nesting': {},
     tailwindcss: {},
     autoprefixer: {},
-  }
+  },
 };
