@@ -1,13 +1,17 @@
-import { FC, useEffect } from 'react';
+import type { FC } from 'react';
+import { useEffect } from 'react';
+import { H1 } from '../heading';
+
+const pageTitle = 'Home';
 
 const Home: FC = () => {
   useEffect(() => {
-    document.title = 'Home';
+    document.title = pageTitle;
   }, []);
 
   return (
     <div>
-      <h1>Home</h1>
+      <H1>{pageTitle}</H1>
     </div>
   );
 };
