@@ -1,41 +1,39 @@
-import { FC, useState } from 'react';
-import { CreatePet } from '../../../api-client/pet';
-import { Link, useNavigate } from 'react-router-dom';
+import type { FC } from 'react';
+import { useEffect } from 'react';
+import { useNavigate } from 'react-router-dom';
 import { HttpError as HttpErrorPartial } from '../../partial/http-error';
 import { PetForm } from '../../form/pet-form';
-import { PetRequest } from '../../../model/model';
-import { HttpError, HttpErrorWithInvalidParameters } from '../../../api-client/error';
+import type { PetRequest } from '../../../../model/pet';
+import { AnchorButton } from '../../button';
+import { H1 } from '../../heading';
+import { useModelResource } from '../../../hook/use-model-resource';
+import { createPetClient as createClient } from '../../../client/pet';
+
+const pageTitle = 'Pet Create';
 
 const Create: FC = () => {
   const navigate = useNavigate();
 
-  const [httpError, setHttpError] = useState<HttpError>();
-
-  document.title = 'Create Pet';
+  const { httpError, actions } = useModelResource({ createClient });
 
   const submitPet = async (pet: PetRequest) => {
-    const response = await CreatePet(pet);
-
-    if (response instanceof HttpError) {
-      setHttpError(response);
-    } else {
-      setHttpError(undefined);
-
+    if (await actions.createModel(pet)) {
       navigate('/pet');
     }
   };
 
+  useEffect(() => {
+    document.title = pageTitle;
+  }, []);
+
   return (
     <div data-testid="page-pet-create">
       {httpError ? <HttpErrorPartial httpError={httpError} /> : null}
-      <h1>Create Pet</h1>
-      <PetForm
-        submitPet={submitPet}
-        error={httpError instanceof HttpErrorWithInvalidParameters ? httpError : undefined}
-      />
-      <Link to="/pet" className="btn-gray">
+      <H1>{pageTitle}</H1>
+      <PetForm httpError={httpError} submitPet={submitPet} />
+      <AnchorButton to="/pet" colorTheme="gray">
         List
-      </Link>
+      </AnchorButton>
     </div>
   );
 };
