@@ -2,10 +2,13 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './app';
 import { StrictMode } from 'react';
+import { BrowserRouter } from 'react-router-dom';
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 root.render(
   <StrictMode>
-    <App />
+    <BrowserRouter future={{ v7_startTransition: true }}>
+      <App />
+    </BrowserRouter>
   </StrictMode>,
 );
