@@ -1,66 +1,73 @@
-import { render, screen } from '@testing-library/react';
-import App from '../src/app';
-import { Router } from 'react-router-dom';
-import userEvent from '@testing-library/user-event';
-import { ReactNode } from 'react';
-import { createMemoryHistory } from 'history';
-import { vi } from 'vitest';
-import { test, expect } from 'vitest';
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+import React from 'react';
+import { test, expect, vi } from 'vitest';
 import { formatHtml } from './formatter';
+import { render, screen } from '@testing-library/react';
+import { MemoryRouter } from 'react-router-dom';
+import App from '../src/app';
+import { userEvent } from '@testing-library/user-event';
 
-vi.mock('react-router-dom', async () => {
+vi.mock('../src/routes', () => {
   return {
-    ...((await vi.importActual('react-router-dom')) as typeof import('react-router-dom')),
-    BrowserRouter: ({ children }: { children: ReactNode; }) => <div>{children}</div>,
+    default: () => <div data-testid="routes-mock" />,
   };
 });
 
-vi.mock('../src/component/page/home', () => {
-  return {
-    default: () => <div data-testid="page-home-mock"></div>,
-  };
-});
-
-vi.mock('../src/component/page/pet/list', () => {
-  return {
-    default: () => <div data-testid="page-pet-list-mock"></div>,
-  };
-});
-
-vi.mock('../src/component/page/pet/create', () => {
-  return {
-    default: () => <div data-testid="page-pet-create-mock"></div>,
-  };
-});
-
-vi.mock('../src/component/page/pet/read', () => {
-  return {
-    default: () => <div data-testid="page-pet-read-mock"></div>,
-  };
-});
-
-vi.mock('../src/component/page/pet/update', () => {
-  return {
-    default: () => <div data-testid="page-pet-update-mock"></div>,
-  };
-});
-
-vi.mock('../src/component/page/not-found', () => {
-  return {
-    default: () => <div data-testid="page-not-found-mock"></div>,
-  };
-});
-
-test('toggle', async () => {
-  const history = createMemoryHistory();
-
+test('close navigation', async () => {
   const { container } = render(
-    <Router location={history.location} navigator={history}>
+    <MemoryRouter>
       <App />
-    </Router>,
+    </MemoryRouter>,
   );
 
-  await screen.findByTestId('page-home-mock');
+  await screen.findByTestId('routes-mock');
+
+  expect(formatHtml(container.outerHTML)).toMatchInlineSnapshot(`
+    "<div>
+      <div class="relative flex min-h-full flex-col md:flex-row">
+        <nav
+          class="absolute flow-root h-16 w-full bg-gray-900 px-4 py-3 text-2xl font-semibold uppercase leading-relaxed text-gray-100"
+        >
+          <button
+            class="float-right block border-2 p-2 md:hidden"
+            data-testid="navigation-toggle"
+          >
+            <span class="block h-2 w-6 border-t-2"></span
+            ><span class="block h-2 w-6 border-t-2"></span
+            ><span class="block h-0 w-6 border-t-2"></span></button
+          ><a aria-current="page" class="hover:text-gray-500 active" href="/"
+            >Petstore</a
+          >
+        </nav>
+        <nav
+          class="mt-16 w-full bg-gray-200 md:block md:w-1/3 lg:w-1/4 xl:w-1/5 hidden"
+        >
+          <ul>
+            <li>
+              <a
+                data-testid="navigation-pet"
+                class="block px-4 py-2 bg-gray-300 ext-gray-900 hover:bg-gray-400"
+                href="/pet"
+                >Pets</a
+              >
+            </li>
+          </ul>
+        </nav>
+        <div class="w-full px-6 py-8 md:w-2/3 lg:w-3/4 xl:w-4/5 mt-16">
+          <div data-testid="routes-mock"></div>
+        </div>
+      </div>
+    </div>
+    "
+  `);
+});
+
+test('open navigation', async () => {
+  const { container } = render(
+    <MemoryRouter>
+      <App />
+    </MemoryRouter>,
+  );
 
   const navigationToggle = await screen.findByTestId('navigation-toggle');
 
@@ -68,86 +75,37 @@ test('toggle', async () => {
 
   expect(formatHtml(container.outerHTML)).toMatchInlineSnapshot(`
     "<div>
-      <div>
-        <div id="wrapper" class="displayMenu">
-          <nav id="top-nav" class="flow-root">
-            <button id="toggle" data-testid="navigation-toggle">
-              <span></span><span></span><span></span></button
-            ><a aria-current="page" class="active" href="/">Petstore</a>
-          </nav>
-          <nav id="left-nav">
-            <ul>
-              <li><a class="" href="/pet">Pets</a></li>
-            </ul>
-          </nav>
-          <div id="main"><div data-testid="page-home-mock"></div></div>
-        </div>
-      </div>
-    </div>
-    "
-  `);
-});
-
-test('home page', async () => {
-  const history = createMemoryHistory();
-
-  const { container } = render(
-    <Router location={history.location} navigator={history}>
-      <App />
-    </Router>,
-  );
-
-  await screen.findByTestId('page-home-mock');
-
-  expect(formatHtml(container.outerHTML)).toMatchInlineSnapshot(`
-    "<div>
-      <div>
-        <div id="wrapper" class="">
-          <nav id="top-nav" class="flow-root">
-            <button id="toggle" data-testid="navigation-toggle">
-              <span></span><span></span><span></span></button
-            ><a aria-current="page" class="active" href="/">Petstore</a>
-          </nav>
-          <nav id="left-nav">
-            <ul>
-              <li><a class="" href="/pet">Pets</a></li>
-            </ul>
-          </nav>
-          <div id="main"><div data-testid="page-home-mock"></div></div>
-        </div>
-      </div>
-    </div>
-    "
-  `);
-});
-
-test('not found', async () => {
-  const history = createMemoryHistory();
-  history.push('/unknown');
-
-  const { container } = render(
-    <Router location={history.location} navigator={history}>
-      <App />
-    </Router>,
-  );
-
-  await screen.findByTestId('page-not-found-mock');
-
-  expect(formatHtml(container.outerHTML)).toMatchInlineSnapshot(`
-    "<div>
-      <div>
-        <div id="wrapper" class="">
-          <nav id="top-nav" class="flow-root">
-            <button id="toggle" data-testid="navigation-toggle">
-              <span></span><span></span><span></span></button
-            ><a class="" href="/">Petstore</a>
-          </nav>
-          <nav id="left-nav">
-            <ul>
-              <li><a class="" href="/pet">Pets</a></li>
-            </ul>
-          </nav>
-          <div id="main"><div data-testid="page-not-found-mock"></div></div>
+      <div class="relative flex min-h-full flex-col md:flex-row">
+        <nav
+          class="absolute flow-root h-16 w-full bg-gray-900 px-4 py-3 text-2xl font-semibold uppercase leading-relaxed text-gray-100"
+        >
+          <button
+            class="float-right block border-2 p-2 md:hidden"
+            data-testid="navigation-toggle"
+          >
+            <span class="block h-2 w-6 border-t-2"></span
+            ><span class="block h-2 w-6 border-t-2"></span
+            ><span class="block h-0 w-6 border-t-2"></span></button
+          ><a aria-current="page" class="hover:text-gray-500 active" href="/"
+            >Petstore</a
+          >
+        </nav>
+        <nav
+          class="mt-16 w-full bg-gray-200 md:block md:w-1/3 lg:w-1/4 xl:w-1/5 block"
+        >
+          <ul>
+            <li>
+              <a
+                data-testid="navigation-pet"
+                class="block px-4 py-2 bg-gray-300 ext-gray-900 hover:bg-gray-400"
+                href="/pet"
+                >Pets</a
+              >
+            </li>
+          </ul>
+        </nav>
+        <div class="w-full px-6 py-8 md:w-2/3 lg:w-3/4 xl:w-4/5 mt-0">
+          <div data-testid="routes-mock"></div>
         </div>
       </div>
     </div>
@@ -156,134 +114,46 @@ test('not found', async () => {
 });
 
 test('pet list', async () => {
-  const history = createMemoryHistory();
-  history.push('/pet');
-
   const { container } = render(
-    <Router location={history.location} navigator={history}>
+    <MemoryRouter initialEntries={['/pet']}>
       <App />
-    </Router>,
+    </MemoryRouter>,
   );
 
-  await screen.findByTestId('page-pet-list-mock');
+  await screen.findByTestId('routes-mock');
 
   expect(formatHtml(container.outerHTML)).toMatchInlineSnapshot(`
     "<div>
-      <div>
-        <div id="wrapper" class="">
-          <nav id="top-nav" class="flow-root">
-            <button id="toggle" data-testid="navigation-toggle">
-              <span></span><span></span><span></span></button
-            ><a class="" href="/">Petstore</a>
-          </nav>
-          <nav id="left-nav">
-            <ul>
-              <li><a aria-current="page" class="active" href="/pet">Pets</a></li>
-            </ul>
-          </nav>
-          <div id="main"><div data-testid="page-pet-list-mock"></div></div>
-        </div>
-      </div>
-    </div>
-    "
-  `);
-});
-
-test('pet create', async () => {
-  const history = createMemoryHistory();
-  history.push('/pet/create');
-
-  const { container } = render(
-    <Router location={history.location} navigator={history}>
-      <App />
-    </Router>,
-  );
-
-  await screen.findByTestId('page-pet-create-mock');
-
-  expect(formatHtml(container.outerHTML)).toMatchInlineSnapshot(`
-    "<div>
-      <div>
-        <div id="wrapper" class="">
-          <nav id="top-nav" class="flow-root">
-            <button id="toggle" data-testid="navigation-toggle">
-              <span></span><span></span><span></span></button
-            ><a class="" href="/">Petstore</a>
-          </nav>
-          <nav id="left-nav">
-            <ul>
-              <li><a aria-current="page" class="active" href="/pet">Pets</a></li>
-            </ul>
-          </nav>
-          <div id="main"><div data-testid="page-pet-create-mock"></div></div>
-        </div>
-      </div>
-    </div>
-    "
-  `);
-});
-
-test('pet read', async () => {
-  const history = createMemoryHistory();
-  history.push('/pet/4d783b77-eb09-4603-b99b-f590b605eaa9');
-
-  const { container } = render(
-    <Router location={history.location} navigator={history}>
-      <App />
-    </Router>,
-  );
-
-  await screen.findByTestId('page-pet-read-mock');
-
-  expect(formatHtml(container.outerHTML)).toMatchInlineSnapshot(`
-    "<div>
-      <div>
-        <div id="wrapper" class="">
-          <nav id="top-nav" class="flow-root">
-            <button id="toggle" data-testid="navigation-toggle">
-              <span></span><span></span><span></span></button
-            ><a class="" href="/">Petstore</a>
-          </nav>
-          <nav id="left-nav">
-            <ul>
-              <li><a aria-current="page" class="active" href="/pet">Pets</a></li>
-            </ul>
-          </nav>
-          <div id="main"><div data-testid="page-pet-read-mock"></div></div>
-        </div>
-      </div>
-    </div>
-    "
-  `);
-});
-
-test('pet update', async () => {
-  const history = createMemoryHistory();
-  history.push('/pet/4d783b77-eb09-4603-b99b-f590b605eaa9/update');
-
-  const { container } = render(
-    <Router location={history.location} navigator={history}>
-      <App />
-    </Router>,
-  );
-
-  await screen.findByTestId('page-pet-update-mock');
-
-  expect(formatHtml(container.outerHTML)).toMatchInlineSnapshot(`
-    "<div>
-      <div>
-        <div id="wrapper" class="">
-          <nav id="top-nav" class="flow-root">
-            <button id="toggle" data-testid="navigation-toggle">
-              <span></span><span></span><span></span></button
-            ><a class="" href="/">Petstore</a>
-          </nav>
-          <nav id="left-nav">
-            <ul>
-              <li><a aria-current="page" class="active" href="/pet">Pets</a></li>
-            </ul>
-          </nav>
-          <div id="main"><div data-testid="page-pet-update-mock"></div></div>
+      <div class="relative flex min-h-full flex-col md:flex-row">
+        <nav
+          class="absolute flow-root h-16 w-full bg-gray-900 px-4 py-3 text-2xl font-semibold uppercase leading-relaxed text-gray-100"
+        >
+          <button
+            class="float-right block border-2 p-2 md:hidden"
+            data-testid="navigation-toggle"
+          >
+            <span class="block h-2 w-6 border-t-2"></span
+            ><span class="block h-2 w-6 border-t-2"></span
+            ><span class="block h-0 w-6 border-t-2"></span></button
+          ><a class="hover:text-gray-500" href="/">Petstore</a>
+        </nav>
+        <nav
+          class="mt-16 w-full bg-gray-200 md:block md:w-1/3 lg:w-1/4 xl:w-1/5 hidden"
+        >
+          <ul>
+            <li>
+              <a
+                data-testid="navigation-pet"
+                aria-current="page"
+                class="block px-4 py-2 text-gray-100 bg-gray-700 hover:bg-gray-600"
+                href="/pet"
+                >Pets</a
+              >
+            </li>
+          </ul>
+        </nav>
+        <div class="w-full px-6 py-8 md:w-2/3 lg:w-3/4 xl:w-4/5 mt-16">
+          <div data-testid="routes-mock"></div>
         </div>
       </div>
     </div>
