@@ -1,22 +1,16 @@
-import { createMemoryHistory } from 'history';
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+import React from 'react';
 import { render } from '@testing-library/react';
-import { Router } from 'react-router-dom';
 import Home from '../../../src/component/page/home';
 import { test, expect } from 'vitest';
 import { formatHtml } from '../../formatter';
 
 test('default', () => {
-  const history = createMemoryHistory();
-
-  const { container } = render(
-    <Router location={history.location} navigator={history}>
-      <Home />
-    </Router>,
-  );
+  const { container } = render(<Home />);
 
   expect(formatHtml(container.outerHTML)).toMatchInlineSnapshot(`
     "<div>
-      <div><h1>Home</h1></div>
+      <div><h1 class="mb-4 border-b pb-2 text-4xl font-black">Home</h1></div>
     </div>
     "
   `);
