@@ -1,8 +1,10 @@
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+import React from 'react';
 import { render } from '@testing-library/react';
 import { HttpError as HttpErrorPartial } from '../../../src/component/partial/http-error';
 import { test, expect } from 'vitest';
 import { formatHtml } from '../../formatter';
-import { HttpError, HttpErrorWithInvalidParameters } from '../../../src/api-client/error';
+import { BadRequestOrUnprocessableEntity, HttpError } from '../../../src/client/error';
 
 test('minimal', () => {
   const httpError = new HttpError({
@@ -13,14 +15,16 @@ test('minimal', () => {
 
   expect(formatHtml(container.outerHTML)).toMatchInlineSnapshot(`
     "<div>
-      <div id="httpError"><p>This is the title</p></div>
+      <div data-testid="http-error" class="mb-6 bg-red-300 px-5 py-4">
+        <p class="font-bold">This is the title</p>
+      </div>
     </div>
     "
   `);
 });
 
 test('maximal', () => {
-  const httpError = new HttpErrorWithInvalidParameters({
+  const httpError = new BadRequestOrUnprocessableEntity({
     title: 'This is the title',
     detail: 'This is the detail',
     instance: 'This is the instance',
@@ -31,8 +35,8 @@ test('maximal', () => {
 
   expect(formatHtml(container.outerHTML)).toMatchInlineSnapshot(`
     "<div>
-      <div id="httpError">
-        <p>This is the title</p>
+      <div data-testid="http-error" class="mb-6 bg-red-300 px-5 py-4">
+        <p class="font-bold">This is the title</p>
         <p>This is the detail</p>
         <p>This is the instance</p>
         <ul>
