@@ -1,24 +1,24 @@
 /// <reference types="vitest" />
 /// <reference types="vite/client" />
 
-// Official Vitest example: https://stackblitz.com/github/vitest-dev/vitest/tree/main/examples/react-testing-lib?file=src%2Futils%2Ftest-utils.tsx
-
-import { defineConfig } from "vitest/config";
-import react from "@vitejs/plugin-react";
+import { defineConfig } from 'vitest/config';
+import react from '@vitejs/plugin-react';
 
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
   test: {
     globals: true,
-    environment: "jsdom",
-    setupFiles: "./vitest.setup.ts",
+    environment: 'jsdom',
+    setupFiles: './vitest.setup.ts',
     include: ['tests/**/*.test.*'],
     coverage: {
       all: true,
       clean: true,
       reporter: ['text', 'html', 'lcov'],
       provider: 'v8',
-    }
+      include: ['src'],
+      exclude: ['src/index.tsx', 'src/vite-env.d.ts'],
+    },
   },
 });
